[2주차] 1단계 UI-로직 분리, FlipCardCore class로 분리
